fix: remove nic finalizer when node deleted
