+++ Dockerfile.controller
@@ -11,6 +11,8 @@ COPY cmd/ cmd/
 COPY api/ api/
 COPY pkg/ pkg/
 COPY controllers/ controllers/
+COPY nodes/ nodes/
+COPY internal/ internal/
 
 RUN CGO_ENABLED=0 GOOS=linux GOARCH=amd64 GO111MODULE=on go build -a -o controller ./cmd/controller/
