+++ cmd/controller/controller.go
@@ -119,6 +119,16 @@ func main() {
 		setupLog.Error(err, "unable to create controller", "controller", "PrivateNetwork")
 		os.Exit(1)
 	}
+	if err = (&controllers.NetworkInterfaceReconciler{
+		Client:      mgr.GetClient(),
+		Log:         ctrl.Log.WithName("controllers").WithName("NetworkInterface"),
+		Scheme:      mgr.GetScheme(),
+		IPAM:        ipam,
+		InstanceAPI: instance.NewAPI(scwClient),
+	}).SetupWithManager(mgr); err != nil {
+		setupLog.Error(err, "unable to create controller", "controller", "NetworkInterface")
+		os.Exit(1)
+	}
 	// +kubebuilder:scaffold:builder
 
 	setupLog.Info("starting manager")
