+++ internal/constants/constants.go
@@ -1,0 +1,15 @@
+package constants
+
+const (
+	// IPFinalizerName is the name of the IP finalizer
+	IPFinalizerName = "scaleway.com/finalizer-ip"
+
+	// FinalizerName is the name of the finalizer
+	FinalizerName = "scaleway.com/finalizer"
+
+	// PrivateNetworkLabel is the private network label
+	PrivateNetworkLabel = "private-network"
+
+	// NodeLabel is the node label
+	NodeLabel = "node"
+)
